% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\title{Parametric bootstrap for location-scale regression}
\usage{
boot(m, num_samples = 1000, seed = NULL)
}
\arguments{
\item{m}{A location-scale regression model from the \code{\link[=lmls]{lmls()}} function.}

\item{num_samples}{The number of bootstrap samples to draw.}

\item{seed}{Either \code{NULL} or an integer that will be used in a call to
\code{\link[=set.seed]{set.seed()}} before simulating the response vectors.}
}
\description{
A simple parametric bootstrap algorithm for location-scale regression models
from the \code{\link[=lmls]{lmls()}} function.
}
