% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lslm.R
\name{lslm}
\alias{lslm}
\title{Location-scale regression}
\usage{
lslm(location, scale = ~1, data = environment(location), light = TRUE)
}
\arguments{
\item{location}{A two-sided formula with the response variable on the LHS
and the predictor for the mean on the RHS.}

\item{scale}{A one-sided formula with the predictor for the standard
deviation on the RHS.}

\item{data}{A data frame (or list or environment) in which to evaluate
the \code{location} and \code{scale} formulas.}

\item{light}{If \code{TRUE}, the design matrices are removed from the estimated
model to save some memory.}
}
\description{
The location-scale regression model assumes a normally distributed response
variable with one linear predictor for the mean (= the location) and one for
the standard deviation (= the scale). The standard deviation is mapped to
the linear predictor through a log link.

This function sets up the model object and estimates it with maximum
likelihood.
}
