% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{MCMC inference for location-scale regression}
\usage{
mcmc(m, nsim = 1000)
}
\arguments{
\item{m}{A location-scale regression model from the \code{\link[=lslm]{lslm()}} function.}

\item{nsim}{The number of MCMC samples to draw.}
}
\description{
A Markov chain Monte Carlo (MCMC) sampler for location-scale regression
models from the \code{\link[=lslm]{lslm()}} function. The sampler uses a Metropolis-Hastings
algorithm with Gaussian proposals that approximate the curvature of the
posterior with the Fisher information. All priors in the model are assumed
to be flat.
}
